moves create project event handler to monitoring package
